fix(cron): address review feedback — envelope rfind, error context, bounds tests, due_jobs filter
Addresses @singlerider's review on #5661.

- parse_shell_envelope: switch `.find` → `.rfind` so the legitimate stdout/stderr
  boundary wins even when stdout contains the literal `\nstderr:\n` sequence.
  Adds `parse_shell_envelope_splits_at_last_stderr_marker` and
  `format_output_for_delivery_handles_embedded_stderr_marker_in_stdout`.

- Delivery validation errors no longer say "for announce mode" for paths that
  aren't `cron announce` (e.g. `cron add --channel X`). Messages now describe
  the field dependency: "delivery.channel is required when configuring channel
  delivery" / "delivery.to is required when delivery.channel is set". Updated
  in both `validate_delivery_config` (cron/mod.rs) and `deliver_if_configured`
  (scheduler.rs) for consistency.

- apply_bounds_after_run now has direct unit coverage: four tests exercising
  remaining_runs decrement-keeps, remaining_runs exhaust-removes,
  expires_until past-removes, expires_until future-keeps.

- due_jobs() and all_overdue_jobs() now filter
  `AND (remaining_runs IS NULL OR remaining_runs > 0)` as a belt-and-suspenders
  safety net for the interrupted-between-decrement-and-remove failure mode.
  Adds `due_jobs_skips_exhausted_remaining_runs`.

- Relocate the previously `#[cfg(all(test, zeroclaw_root_crate))]`-gated cron
  test module from `crates/zeroclaw-runtime/src/cron/mod.rs` to
  `src/cron/mod.rs`, where `handle_command` and `CronCommands` actually live.
  The cfg gate is now a plain `#[cfg(test)]`, so these 26 CLI-surface tests
  finally execute in CI. Added `count: None, until: None` to six
  `CronCommands::Add` fixtures that predated the bounds feature.

Validation on aarch64 Pi 5:
- `cargo fmt --all -- --check` clean.
- `cargo clippy -p zeroclaw-runtime -p zeroclawlabs --all-targets -- -D warnings` clean.
- `cargo test -p zeroclaw-runtime cron` → 154 pass / 0 fail (was 153; +7 new).
- `cargo test -p zeroclawlabs cron::tests` → 26 pass / 0 fail (were dead code).
